backend/requirements.txt
# Smart Find — FastAPI + OCR + optional GPT-4 / Donut (see requirements-ml.txt)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pymupdf>=1.23.0
rapidfuzz>=3.6.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.26.0
easyocr>=1.7.0
openai>=1.30.0
